Updated the priorities in the README
Indexed quite a few of the fields in entries table, to speed up lookup and location retrieval.

Added proximity to the entries table and made some methods to add this value to all entries. This is required to determine and store the number of entries nearby each entry.

Renamed fields lng and lat to longitude and latitude, because Geocoder wasn't allowing otherwise. Updated all the specs to reflect this change