Added two migrations to expand the integer size the tags table could capture.
Added the .hoover_tag class method to Entry, allowing me to grab the whole backrecord of entries associated with a specific tag. This frees me to drop/reset the database more readily.

Added a new rake task for hooving a tag. Removed the prox task, as it is now redundant.

Added a test to ensure that tags were being saved appropriately, since thay were not previously.