Moved some logic out of the feed action into the controller (Moving more next commit)
Changed the maptype to ROADMAP so that you can zoom in a lot further

Added a configuration for GEOCODER to the initializers folder. This is required to avoid rate-limiting on Google (which is the default).

Tweaked the cron-jobs, not sure if 60+ minutes is a valid input

Reinserted the db:prox rake task, to update the proxmity count of entries