Added styling for device-location button.
Added event listener for device-location button.

Created customLeaflet.deviceLocation event listener for the device-location button, in order to trigger a function when the user wants to allow the use of device location.

Created customLeaflet.onLocationFound to act as callback for successful locate event

Created customLeaflet.roundedLatLng to prepare any latlng object for passing to the dataRequest function

Updated cron schedule