Built out the onClick function for markers. Now they integrate popups, enlarging the icon, pulling the icon to the front and can be clicked on and off.
This required an extra icon constructor, woop.

The currentOffset shenanigans is there to ensure if there are multiple markers that have been enlarged, the most recently enlarged marker will come to the top of the markers.

Passing the thumb url as one of the marker options, so I can access it once the click event occurs

Bound a popup to each marker, with the full url of the entry, so that users can check those out.