Added styling to mapbox tiles, and thus changed the tile URL to fetch the new tiles.
Changed the zIndexOffset to MAKE SURE the clicked entry would be shown over all the others.

Ensured that the presentMarker function would work for markers in a cluster (with a zIndexOffset of 1000000) and markers by themselves (with a zIndexOffset of 0)

Added spiderify radius option to the marker cluster, so that it spaces the markers out some more. Make play with this a little more.

Added basic legend to map, with stylings.

Added simple black border to all marker images.

Messed with the gradient of the heatmap again....